saem: address Copilot review on PR #30
- get_mu_ref_pairs: exclude additive mu-refs (log_transformed=false). The
  gradient-step chain rule applied to log_theta only holds for lognormal
  forms; additive mu-refs now fall through to the regular NLopt path.
- theta_sigma_mstep_light: skip central-FD evaluations for pinned dims
  (lower==upper). Without this the gradient-step branch couldn't actually
  save NLopt OFV calls because central FD still hit the pinned mu-ref dims.
- Replace the hand-wavy `count * pairs * 3` evals-saved formula with a
  per-call accumulator (`2 * mstep_maxiter * pairs`) that reflects the
  real upper bound on skipped obs_nll_sum calls. Field bumped to u64.
- Add unit tests for get_mu_ref_pairs covering the empty case, log/additive
  filtering, and orphaned-theta skip.

+++ src/estimation/outer_optimizer.rs
@@ -18,7 +18,7 @@ pub struct OuterResult {
     pub warnings: Vec<String>,
     /// Estimated OFV evaluations saved by the SAEM mu-ref gradient step M-step.
     /// Non-None only when method=saem and mu_referencing=true.
-    pub saem_mu_ref_m_step_evals_saved: Option<u32>,
+    pub saem_mu_ref_m_step_evals_saved: Option<u64>,
     pub ebe_convergence_warnings: u32,
     pub max_unconverged_subjects: u32,
     pub total_ebe_fallbacks: u32,
